Feature/move asset profile icon component to @ghostfolio/ui (#4858)
* Move asset profile icon component to @ghostfolio/ui

* Update changelog
