+++ apps/client/src/app/pages/accounts/transfer-balance/transfer-balance-dialog.module.ts
@@ -1,4 +1,4 @@
-import { GfAssetProfileIconComponent } from '@ghostfolio/client/components/asset-profile-icon/asset-profile-icon.component';
+import { GfAssetProfileIconComponent } from '@ghostfolio/ui/asset-profile-icon';
 
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
