+++ libs/ui/src/lib/asset-profile-icon/asset-profile-icon.component.ts
