+++ libs/ui/src/lib/asset-profile-icon/index.ts
@@ -1,0 +1,1 @@
+export * from './asset-profile-icon.component';
